tasklane-ai-pipeline/requirements.txt
torch
torchvision
torchaudio
whisperx
opencv-python
ffmpeg-python
paddleocr
ultralytics
segment-anything
langchain
openai
python-dotenv
rich
numpy
Pillow 